perf(cli): keep the dev watcher off the backend-core command graph
PERF-100. The long-lived `kitcn dev` watcher child imported all 7,828 lines of
backend-core for three symbols, so it booted execa and the whole command graph
and held it for the session. Move `resolveConfiguredBackend` to `config.ts` and
`withLocalCodegenEnv` to a new `local-env.ts`, resolve `generateMeta` and
`loadCliConfig` directly instead of through `resolveRunDeps`, and load execa
only on the concave branch that actually shells out.
